fix recursive loop
