--- conf_sp_idp/sp/app/models.py
@@ -1,42 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-
-    conf_sp_idp.sp.models.py
-    ~~~~~~~~~~~~~~~~~~~~~~~~
-
-    Models for the simple service provider (sp)
-
-    This file is part of MSS.
-
-    :copyright: Copyright 2023 Nilupul Manodya
-    :copyright: Copyright 2023 by the MSS team, see AUTHORS.
-    :license: APACHE-2.0, see LICENSE for details.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-"""
-
-from flask_login import UserMixin
-from conf_sp_idp.sp.app.app import db
-
-class User(UserMixin, db.Model):
-    """Class representing a user"""
-
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(250), index=True, unique=True)
-
-    def __repr__(self):
-        return f'<User {format(self.email)}>'
-
-    def get_id(self):
-        """Get the user's ID"""
-        return self.id
